+++ importer/management/commands/import_tables.py
@@ -8,7 +8,7 @@ from clinvar.models import Clinvar
 from conservation.models import KnowngeneAA
 from dbsnp.models import Dbsnp
 from frequencies.models import Exac, GnomadExomes, GnomadGenomes, ThousandGenomes
-from geneinfo.models import Hgnc, Mim2geneMedgen, Hpo, NcbiGeneInfo, NcbiGeneRif
+from geneinfo.models import Hgnc, Mim2geneMedgen, Hpo, NcbiGeneInfo, NcbiGeneRif, HpoName
 from hgmd.models import HgmdPublicLocus
 from ...models import ImportInfo
 from pathways.models import EnsemblToKegg, RefseqToKegg, KeggInfo
@@ -25,7 +25,7 @@ TABLES = {
     "gnomAD_genomes": (GnomadGenomes,),
     "hgmd_public": (HgmdPublicLocus,),
     "hgnc": (Hgnc,),
-    "hpo": (Hpo,),
+    "hpo": (Hpo, HpoName),
     "kegg": (KeggInfo, EnsemblToKegg, RefseqToKegg),
     "knowngeneaa": (KnowngeneAA,),
     "ncbi_gene": (NcbiGeneInfo, NcbiGeneRif),